:bug: Fixing header arguments bug
During the implementation of some cronjob triggers we've discovered that
Kubeless controller is not parsing our event data.

To fix that we've changed the patter of the headers arguments on curl command to
be capitalized. So, event-time is not sent as Event-Time.

Now, we can access event data on cronjob triggers.