Feature: Add payload option to trigger (#12)
* :bug: Fixing header arguments bug

During the implementation of some cronjob triggers we've discovered that
Kubeless controller is not parsing our event data.

To fix that we've changed the patter of the headers arguments on curl command to
be capitalized. So, event-time is not sent as Event-Time.

Now, we can access event data on cronjob triggers.

* :bug: Fixes curl static args

After some testing with our fix, I've discovered that the some args are static.
The event-id and event-time, for example, are always the same for every execution.

On this commit I've fixed that, changing the way we pass those args.

Now, we have dynamic args.

* :art: Changing from JOB_NAME to POD_UID on event-id

After talking with Kubeless maintener, we decided to change the ID we're using
for the event. Instead of the pod name, we're going to use the metada.uuid.

On this commit I've changed that.

Now, we have a properly unique event-id.

* :bug: Fixing automated tests

During the PR creation I've found out some issues on our automated tests.

On this commit I've updated our tests and fixed thos issues.

Now, we can merge our branch.

* :art: Applying lint

* :sparkles: Adds payload support to controller

During the implementation of Kubeless in our infrastructure, we decided to
create a generic function that receives arguments from a cronjob.

On this commit I'm adding a way to add payload data to CronJob triggers.

Now we can pass data to the `CronJobTrigger.Spec.Payload` item.

* :bug: Fixes lint errors

After the creation of the PR I've found out some lint errors.

On this commit I'm fixing those errors.

Now our code should be following the desired patterns.

* :art: Applying PR suggestions