:sparkles: Merges trigger and function metadata
During the implementation of the CronJob triggers on QuintoAndar, we found out
that Istio doesn't handle nicely with jobs. Specially our cronjobs, because
they're too fast (a simple curl) and Istio can't create the proxy before our
container ran the command.

On this commit we're merging Labels and Annotations, and passing them to our
jobs. By doing so we can now add the sidecar removal annotation to our
CronJobTrigger and it will pass down to the Job itself.

Now, Istio can be disabled on our CronJobs.