:bug: Fixing job name
BLAC-1700

Our controller is manually creating the job name based on the function name. Because of that, it is impossible to deploy multiple triggers for the same function.

On this commit, I've moved from creating a new name for our trigger to using the CronJobTrigger resource name instead.

Now, we can create multiple cron triggers for the same function.